Backend/NetLabBuilder/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docker==6.1.3
idna==3.4
ipaddress==1.0.23
packaging==23.1
pathlib==1.0.1
requests==2.31.0
urllib3==2.0.4
websocket-client==1.6.2


flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
docker==6.1.3
gunicorn==20.1.0
psutil==5.9.5
mysql-connector-python==8.2.0
scapy==2.5.0
